set/getrlimit, set/getuid and setlogmask bindings
Hi,

  while packagin the jabber server "prosody" for Debian, I noticed that
it was using a custom binding of some posix function not available in
luaposix. After asking prosody author, I wrote a patch for luaposix
that adds the missing functionalities prosody is actually shipping in a
custom .so.

The attached patch binds the following functions:
set/getrlimit, set/getuid and setlogmask.

I'd like it to be integrated in a future release of luaposix, if
posible.

Cheers
--
Enrico Tassi