Fix compilation and runtime on QNX based systems
This patch updates the configuration and implementation allow this
product to compile and build on QNX based platforms. QNX should be
treated like darwin and openbsd in terms of posix compliance, however it
does not support the SystemV msg* interface to message queues.  It does
support the mqueue interface but implementing that is beyond the scope
of this patch.