socket: handle AF_UNIX socket path string lengths properly.
Close #232
* ext/include/_helpers.c (checkstringfield): Add a length pointer,
and...
(checklstringfield): ...rename to this.
(checkstringfield): Add a macro wrapper for the old API.
(pushlstringfield): New macro to push an lstring.
* ext/posix/sys/socket.c (pushsockaddrinfo): Use pushlstringfield
for AF_UNIX path field so as not to terminate at the first \0.
(sockaddr_from_lua): Use memcpy for the same reason.
* examples/abstract-ns-server.lua, examples/abstract-ns-client.lua:
New examples demonstrating non-NULL terminated AF_UNIX paths to
us the Linux abstract namespace.
* NEWS.md (New Features, Bugs Fixed): Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>