Allow users to use GLOB_MARK flag with glob()
Most of the optional flags for glob wouldn't make much sense for Lua,
but GLOB_MARK is very useful and would require a lot of work for a user
of luaposix to add after the fact. (When GLOB_MARK is added to a call of
glob, directories are automatically given a final '/' in the list of
return values.)

The function signature changes from glob([pattern]) to glob([pattern],
[glob_mark]). Any truthy value passed as a second argument turns on
GLOB_MARK. The second argument is optional, and the flag defaults to
false. So users who don't want this extra can simply ignore it, and they
can contingue to call the function as glob() and get the shortcut for
the "*" pattern.

I haven't added tests yet. If there's interest in this patch, I'll spend
a little time learning specl and do so. (Looks like tests for glob
should go in spects/posix_spec.yaml, yes?)