Fix time functions for platforms with tm_gmtoff / tm_zone.
POSIX only guarantees that struct tm has a minimum set of members;
implementations may add more. Commonly-used implementations (such as
the IANA tzcode libraries and their derivatives) add members
tm_gmtoff, with the seconds offset from UTC, and tm_zone, with a
pointer to a timezone abbreviation.

strftime() depends on these fields having been filled in by gmtime()
or localtime(), which means that we must preserve them if they exist
when passing the struct tm to or from Lua.

Failure to do so results in nonsense results from the %z format, and
may cause crashes with the %Z format.