Try and make tests more portable.
POSIX doesn't define the "execute" permissions bits for IPC objects,
and so setting them isn't portable.

POSIX defines the group owner of a new file as being EITHER the egid
of the creating process OR the group of the parent directory (BSD
behavior). The test suite was assuming the former. Avoid the issue by
changing the group of the parent dir to the egid (which should be
allowed even under restricted chown).

Don't assume that IPv6 sockets can send to IPv4 addresses, because
nothing says they should be able to.

Squish whitespace in the default uname string because for some reason
that is failing due to whitespace differences.

This doesn't quite make everything pass under FreeBSD but it gets very
close. Of the remaining errors I get, one is a documented spec
violation (nice() returning 0 rather than the niceness), one is an
unrelated luaposix bug (integer range of rlim_t values), and one is
still unresolved.