socket.c: fix getsockname() and getpeername()
Both take addrlen as input parameter, which is now properly initialized
to the size of the sockaddr_storage and not some
random value from the stack.

The returned path in sockaddr_un is \0 terminated. Use memchr(sun_path)
to calculate the real sun_path length instead of 108 (sizeof sun_path).
Otherwise we get a lua string with embedded zeros and other garbage.

Signed-off-by: Christian Hohnstaedt <christian@hohnstaedt.de>